:recycle: migrate from SSH to SSM and restructuring
